Move structures definitions to separate header
